requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
pymongo==3.12.0
gspread==5.1.1
google-auth==2.3.3
python-dotenv==0.19.0
gunicorn==20.1.0
dnspython==2.1.0